Bug 1287234 - Make finding remote mozreview repository a little more robust. r=gps
When for some reason the output of git cinnabar hg2git doesn't fit the
"one sha1 per line" model, the current code can go horribly wrong by
splitting words instead of lines, assuming that anything that is not a
0000000000000000000000000000000000000000 is a valid sha1, assuming the
output has the same number of lines as the input and still trying other
remotes even when one was already found.

This change makes it all more robust.

MozReview-Commit-ID: 79iHG2qgInO