robustcheckout: resolve symbolic revisions from remote (bug 1287506); r=Callek
Previously, if we were checkout out a symbolic revision, we would
always pull from the remote then resolve that name locally. This
is not ideal because the local repo may have other references to
that name not on the remote and the local references may be checked
out.

We change the code to resolve the name on the remote then use the
resolved value locally. As part of this, we refactor the code to
avoid pulling if the remotely resolved name is locally available.
In theory, we may not pull things like phases and obsolescence
markers that could have been updated on the remote. However, this
is an existing deficiency in the robustcheckout extension. If we
want to ensure all remote state is present locally, we should add
an --always-pull flag or something. But that's for another bug.

MozReview-Commit-ID: 4dSEZITAL2D