hgmo: handle obsolete changesets in automationrelevance endpoint
The json-automationrelevance endpoint is queried by automation to
obtain information on changesets relevant to the current revision
being built.

Because automation may not start immediately after a push was made,
automation may query a changeset that has since become hidden.
For most APIs, this would issue a 404 because hidden changesets are,
well, hidden and as good as not existing. However, for hidden
changesets referenced by the pushlog, we've decided to expose
their information to automation so it can make intelligent
decisions.

This commit allows the automation relevance endpoint to return
information about hidden changesets.

MozReview-Commit-ID: 3JmQT3LQU3p