mozreview: Tag BMO comments when posting reviews and replies (bug 1226080); r=glob
The Bugzilla.post_comments() method now uses the MozReview.attachments() BMO
API, which allows tagging comments with mozreview-* tags even if the current
user is not a member of the editbugs group, which is usually required for
comment tags.

A side effect is that comments from reviews and replies will always be
associated with the corresponding MozReview attachment, which is more
appropriate than posting independent comments.  This is why the logic
around parent reviews had to change in an earlier commit to be associated
only with the bug of the first commit (and is now assoicated with that
attachment).

MozReview-Commit-ID: D6ERzdLVPe9