pushlog: tie recordpushes() to a repo transaction (bug 1288859); r=glandium
Before, recordpushes() always committed the database transaction. If
we called recordpushes() inside a Mercurial transaction and rolled back
Mercurial, the pushlog changes would persist. That's not desirable.

We now tie the pushlog transaction to the Mercurial transaction, just
like is done during recordpush().

We require a transaction to be passed into the function because every
caller of this API in tree failed to find the transaction via
self._transref(). I suspect repo filtering foo. I don't think there
are any out-of-tree consumers of this function, so the API change
should be acceptable.

MozReview-Commit-ID: 9FkXYPJe143