testing: vagrant provisioning for dev environments (bug 1294688) r=glob,gps,smacleod
Added a Vagrantfile and Ansible provisioning scripts for the
version-control-tools development environment under a virtualbox guest.

The Vagrantfile requires the 'nugrant' plugin, which allows developers to
override default guest settings like CPU count, memory, and provisioning script.

This is proof-of-concept code, it requires testing by others.  We can include
it in the general developer documentation once it's proven stable and useful.
I've included basic documentation in the Vagrantfile and configure.yml file
bodies.

MozReview-Commit-ID: JaWC5j4woFw