autoland: don't reference an undefined variable in the error case (Bug 1294729). r=glob
cmd_output was being referenced in the exception handler, but when
run_hg() would fail it is never assigned. This was causing autoland
to repeatedly loop on a single changeset due to how it would handle
this unexpected exception.

We should no longer get stuck on a single revision when the related
invocation of run_hg() fails.

MozReview-Commit-ID: 5m3aoRMtKA5