configwizard: tweak tests for multiple vct checkouts; r=glob
The configwizard extension itself counts as an extension running from a
v-c-t checkout. Therefore, we only need a single extension in the test
outside of $TESTDIR.

We add an explicit test for running extensions from $TESTDIR to verify
we get no false positives. We also remove an extra extensions from
$TESTTMP to simplify the test.

MozReview-Commit-ID: lHdXUeiPCx