hghooks: add hook to announce pending retirement of fx-team (bug 1296392); r=jgriffin
This commit contains a simple hook that prints a message about the
retirement of the fx-team repo. The hook is non-invasive and doesn't
(yet) reject pushes.

I plan to install this hook on the integration/fx-team repo on
hg.mozilla.org ASAP.

r+ indicates sign-off of deploying the hook to production.

MozReview-Commit-ID: BFpJCEA30io