docker: use user-defined network for hgmo cluster
User-defined networks are how you are supposed to do networking in
modern versions of Docker. So switch to them.

Gone are the old style links. Instead, you can define aliases in the
networking configuration and these magically appear as DNS entries.
This means hostnames and ports inside containers are static (yay)!
This means the checks at container start time for certain environment
variables can go away. Of course, the downside is if the container
is started in isolation we can't detect this as proactively. Such is
the Docker way, I suppose.

MozReview-Commit-ID: AaCXd3TwhRC