docker: restore streamlined starting of hgmo cluster
A few commits ago, we changed the ordering of the container create+start
sequence because we needed to have some container names in order to create
links at subsequent container create time. Now that we use a
user-defined network and network alias names are set, this isn't needed.
So restore the previous ordering.

MozReview-Commit-ID: ILnL04IflzQ