vcsreplicator: explicitly record obsolescence markers in pushkey messages
Before, obsolescence markers were only recorded during some
transactions, notably if the pushkey protocol was used to set them.

Work in bug 1288845 to allow commit dropping to be performed remotely
on the server revealed that obsolescence markers were not being written
from server local operations.

This commit adds explicit logic to handle obsolescence markers during
transactions. Essentially, we look at hook arguments to see the
number of obolescence markers added by the transaction, obtain the
last N markers, then construct a pushkey message with their serialized
content. This ensures that all created obsolescence markers are
synchronized, even if the pushkey protocol isn't used.

As part of this change, we update the test to run the local consumer
as the mirror user. This is necessary because the server is using
the obsolescencehacks extension to allow certain users to see an
unfiltered repo, which is necessary for clients to pull down
non-hidden heads. The test for an out-of-sync client applying
replication changes involving obsolescence markers triggers this
need.

MozReview-Commit-ID: KTQWRWRYn1O