vcsreplicator: send Pulse messages with obsolescence info (bug 1296368)
Downstream repository watchers may want to know when a changeset is
obsoleted. For example, Firefox automation may wish to cancel
queued and running tasks related to a now-defunct changeset. In
the case of the autoland repo, which will obsolete changesets
instead of dropping them, we want an out-of-band mechanism to inform
the author of the dropped changeset that it has become obsolete
(this mechanism is likely pulsebot).

This commit introduces handling of "obsolete" pushkey messages to
the notifications system. We now decode metadata in obsolescence
markers, convert it to something useful to downstream consumers,
and publish the results to Pulse.

The feature is still experimental and its API may change in the near
future.

MozReview-Commit-ID: C3MhZToETti