testing: pin BMO commit to work around environment bustage (bug 1298524)
docker/generate_bmo_data.pl was removed by the commit after this pinned
commit, breaking our Docker environment. As a temporary stop-gap to fix
the environment, pin the commit.

As part of this, we had to revert back to a non-shallow clone otherwise
there is no guarantee the Git repo has the commit in question. I didn't
want to bloat scope to change `git fetch` to "unshallow" the local
clone.