hgmo: eliminate race condition on kafka topic availability
Before, there was a race condition in the hgmo cluster between creating
a Kafka topic and that topic becoming usable. This resulted in
intermittent "UnknownTopicOrPartition" errors, especially when running
many tests concurrently.

We add code to cluster startup to wait for topic metadata to become
available. This appears to be sufficient to eliminate the race
condition.

With this change, I no longer see UnknownTopicOrPartition errors when
running 6 tests concurrently. Before, I used to get 2+ failures per
test run.

MozReview-Commit-ID: ExPzMAiM3Mq