mozreview: don't show stack when pushing merges (bug 1298992); r=glandium
It appears the behavior of git-cinnabar or git-mozreview changed in the
past few months with regards to error handling during
`git mozreview push`. Today, Python code in git-cinnabar raises an
uncaught exception in `cinnabar-debug-push.py`. This causes Python to
print the traceback and exit 1. This output is then printed to the user.

That is arguably an acceptable outcome for legitimate bugs. However,
the known case of pushing merges results in the traceback, and that is
a poor user experience.

We add code to detect the exception raised from pushing merges and
convert it to an explicit exit code which is gracefully turned into an
abort without displaying the traceback. Sniffing exception messages
isn't a great way to detect error type. But cinnabar lives us no choice
since it raises a generic Exception (and who can fault it - you
shouldn't have to consume its internal API).

MozReview-Commit-ID: 8foHdIHWOFs