docker: ability to undo creation events when an exception occurs (bug 1299337); r=glob
It is (unfortunately) somewhat common to hit errors when starting
Docker clusters. A common trigger for errors appears to be Docker
not scaling well to concurrent requests (e.g. when starting 6
MozReview clusters at a time).

Previously, if we encountered a Docker error during startup, created
containers and networks may linger until test execution has finished.
This could cause unacceptable test overhead or interfere with
subsequent tests.

This commit introduces a context manager that effectively proxies a
docker.Client instance and monitors for created containers and
networks. If an exception occurs when the context manager is active,
created containers and networks are deleted automatically.

We should eventually use this context manager elsewhere. For now, I'm
starting with the MozReview cluster start because that's where most
failures appear to be.

MozReview-Commit-ID: Dy2CZ7Jfmj8