testing: require HTTP 200 response code for HTTP container to be alive (bug 1299337); r=glob
Before, HTTP 500 was returned by the BMO container before the HTTP
endpoint was able to establish a connection to MySQL. It seems like
a good idea to require the monitored URL to issue a HTTP 200 response
for us to assume the container is healthy.

This change significantly reduces the intermittent failure/skip rate
of MozReview tests on my machine. Before, nearly every test would be
skipped if I cranked up the concurrency to say -j6. Now, maybe only
25% of tests are skipped.

MozReview-Commit-ID: 7HZr0llfASS