mozreview: allow reviewing a commit message by injecting it as a FileDiff (bug 1248008) r=smacleod
Currently there is no way to comment on a commit message, reviewers are using bogus diff comments. This patch is injecting the contents of the commit message as a FileDiff.

We're using a standard FileDiff. It appears only in ReviewBoard. In every other aspect it works as a file from repository. As files need to be unique it is named "commit-message" with a suffix (4 characters sliced from mercurial hashtag). The suffix is then removed from UI using a CSS hack and JavaScript. There is a `mr:diff-context-loaded` event added to ReviewBoard fork and a corresponding listener in MozReview to change the filename in `#diff_index` list.

Please note that the changes doesn't affect existing ReviewRequests. No commit message FileDiff is added in that case.

MozReview-Commit-ID: JsBXhGyzUmz