hgmo: change test for json-filelog to not massage data (bug 1300061)
Before, the test was inserting a "pushdate" key in the JSON with a
static value. When we accidentally dropped "pushdate" from the JSON
in db73b5e54c1d, this regression was not detected because of this
test's behavior.

Other tests pave over pushdate value inconsistency by performing a
regex match over the values.

This commit changes the test to behave like others and not massage
the response body before printing. The diff is a little gnarly because
of whitespace differences and key sorting. But we can see that
"pushdate" no longer exists.