configwizard: use [hostsecurity] section to pin fingerprints (bug 1304793); r=smacleod
Mercurial 3.9 supports a new [hostsecurity] section that is everything
[hostfingerprints] was and more.

We add code to detect when running on Mercurial 3.9+. If so, we set
fingerprints in [hostsecurity]. Fingerprints are defined as SHA-256,
since that is harder to create a hash collision for than SHA-1.

If [hostfingerprints] is present, we port fingerprints to
[hostsecurity] automatically.

It's worth noting that we never delete pinned fingerprints. However,
fresh runs on Mercurial 3.9+ with modern SSL won't pin the
fingerprints.


MozReview-Commit-ID: 3HRNy1Wbg5x