mozreview: Improve the error message when pushing to a submitted/discarded review request (bug 1240725) r=smacleod
In addition to simplifying the error message (removing the term 'parent', which
isn't exposed by the UI), also provide instructions on what the user needs to
do to fix the issue.

Switch the error from a FORM_DATA exception to the better suited
REVIEW_REQUEST_UPDATE_NOT_ALLOWED, which further removes cruft from the error
message displayed to the user.

MozReview-Commit-ID: 8ttNIGwNL51