autoland: switch requirements.txt style and pin all packages (Bug 1311066). r=gps
The release of pycparser 2.15 broke autoland. Very few of the python
packages for autoland were pinned so I've switched the requirements.txt
to the new style and pinned all packages with hashes.

Since this change pycparser 2.16 has been released which might fix
things, but that can be looked at in a different change.

MozReview-Commit-ID: JSmVSDblKtp