github-webhooks: AWS Lambda support for consuming GitHub web hooks (bug 1170600); r=dividehex
This commit contains AWS Lambda functions for ingesting GitHub Web Hooks
into Amazon Firehose+SNS, and to republish events from SNS to Pulse.

It also contains Python code for deploying the code to Lambda.

The AWS infrastructure configuration lives in Terraform in a separate
repository. We do use some AWS APIs as part of the deploy that could
be implemented in Terraform. However, their use is minimal. The code
in this repository is narrlowly tailored to the application, run-time
level, not the infrastructure level. So I think it is acceptable.

MozReview-Commit-ID: GiI4Tj3tcQL