hgmo: update moz.build evaluation sandbox to work on CentOS 7 (bug 1263973); r=kang
When we upgraded hg.mozilla.org from CentOS 6 to CentOS 7, the
moz.build evaluation sandbox stopped working. This commit is
an attempt to port it to CentOS 7.

One of the changes in CentOS 7 is many mounts are now marked as
"shared" by default. Even though the child process has a new mounts
namespace via CLONE_NS, unmounts in the child process were being
propagated to the parent! This was resolved by recursively
remounting / with MS_SLAVE to disassociate all mounts from the
parent.

CentOS 7 also introduced the /run/user/1000 mount, which is a tmpfs
volume. We updated the unmounting code to handle this.

MozReview-Commit-ID: CtFqoiq5qx6