ansible/hg-web: allow mozbuild-eval to set its cgroup (bug 1263973); r=kang
See the inline comment for why we need this.

This "just worked" in CentOS 6. My guess is the default permissions
for cgroups changed.

The other alternative to this is making the tasks file permission 777
so anybody can add tasks to it. If there aren't security risks with
that, that approach is more flexible because it doesn't limit the users
that can run `mozbuild-eval`.

MozReview-Commit-ID: D7WD1CS89pj