autoland: don't pass unicode values into transplant (bug 1315253); r=smacleod
Production autoland got into an infinite loop because
formulate_hg_error() in transplant.py was raising a UnicodeDecodeError
trying to coerce str and unicode types because one of the command
arguments was unicode. The unicode instance appeared to originate with
the database.

This commit coerces some values from the database back to str to avoid
the type matching. For good measure, we also add some asserts to
transplant() to guard against unwanted unicode types.

MozReview-Commit-ID: 9YgvS6OSDAr