hgmo: add config option to control and change default behavior of automationrelevant (bug 1297162); r=smacleod
The automationrelevant() revset determines what changesets are relevant
to automation given an input changeset(s). It essentially allows
automation to see what file changes should influence scheduling.

At the time it was implemented, the only non-publishing repository
that automation ran against was Try. Try is special in that heads
on Try are often pushed on top of previous non-public heads. We want
scheduling on Try to be influenced by non-public changesets from
previous pushes because "unfinished" work should continue to
influence scheduling.

The "autoland" repository has since been introduced and it is also
non-publishing. However, draft ancestors from previous pushes
are effectively "finished" and not relevant to automation scheduling.
The behavior of the revset to union with draft ancestors meant that
scheduling on "autoland" frequently brought in file changes from
dozens of draft ancestors from previous pushes that shouldn't have
been influencing scheduling on this repository.

This commit introduces a config option for controlling the behavior
of adding draft ancestor changesets to the "automation relevant" set
so both Try and "autoland" behave the way they should.

We may introduce automation to more non-publishing repositories in
the future. My guess is these repositories will behave more like
"autoland" than Try with regards to the automation relevance of
draft ancestors. Therefore, this commit changes the default behavior
to *not* union with draft ancestors. This means we'll have to update
a config option on the Try repository to maintain current behavior.

MozReview-Commit-ID: 9O6DmUqPhF3