vcsreplicator: teach consumer to wait for N messages and no-op
A number of tests were explicitly waiting on various events in
various processes in order to avoid race conditions in message
delivery. There was significant overhead and complexity to this.
On top of that, it wasn't robust: there were explicit sleeps to
wait for messages to send!

In this commit, we introduce a `--wait-for-n <int>` flag to the
replication consumer. It waits until N messages are available
then exits.

Using this new feature, we change a number of tests to wait for
messages to become available before proceeding. This avoids
race conditions and makes tests simpler.

MozReview-Commit-ID: AcvSu25CzTr