vcsreplicator: publish events to SNS (bug 1316952); r=fubar,glob
We want to publish events to SNS to make it easier for infrastructure
in AWS to react to them. (With Pulse, you have to run a server to
listen for Pulse events. With SNS, you can e.g. register Lambda
functions to react to events so you don't have to run a server.)

This commit introduces a SNS notification daemon. It is very similar
to the existing Pulse one. It simply republishes received events on an
SNS topic.

Ansible configuration for the new daemon has been added. This required
a systemd service, configuration file entries, support for grabbing
secret credentials from a file during deploy, and a few misc other
things.

As usual, writing tests for this functionality proved to take more
time than implementing the feature itself.

MozReview-Commit-ID: AuKWwD7UBal