mozreviewpulse: absorb needed mozreviewbots libraries and delete the rest (Bug 1287537). r=mars
mozreviewbots is no longer going to be integrated into MozReview.
Instead automation will be triggered elsewhere and results will be
consumed from pulse.

mozreviewbots has some useful code for posting reviews to Review Board
in a single request which we'll pull into mozreviewpulse (It could
be argued this should be its own stand-alone library, but we'll keep
it in mozreviewpulse now and possibly pull it out later).

MozReview-Commit-ID: 1qxr8YNzP4h