hgext: remove pushrebase extension; r=smacleod
This was added a while ago so hg.mo could rebase changesets
automagically without requiring the client to rebase. Autoland handles
the use case of pushrebase and I'm pretty sure nobody uses this
functionality because nobody has send anything about it to me. Since
the code breaks in Mercurial 4.0, let's just remove it.

MozReview-Commit-ID: HMWdpXC6aSb