mozautomation: function for hyperlinking a commit message (bug 1325782); r=wlach
hg.mozilla.org currently has a string filter for hyperlinking Bugzilla
bugs. It basically calls commitparser.BUG_RE.sub() to replace bug strings
with `<a href="...`.

As part of vendoring Servo in mozilla-central, we wish to hyperlink GitHub
pull requests and issues. This requires commit message rewriting more
advanced than a simple regular expression substitution.

The existing BUG_RE matches sequences of digits and "#\d+". For a commit
message authored for GitHub, "#\d+" almost certainly refers to a GitHub
pull request or issue, not a Bugzilla bug. So, we can't use BUG_RE for
GitHub authored commits.

This commit implements a new function for hyperlinking a commit message.
It first looks for our source repository and source commit annotations
that our GitHub commit importer (being designed in bug 1322769) adds to
commit messages. It links the source repository if it is a URL.

If the source repository is github.com, we also link to the original
commit for that GitHub project. That's pretty useful for archeologists.

For GitHub sources, we also link "#\d+" to the GitHub issue for that
project. Per https://help.github.com/articles/autolinked-references-and-urls/
I think this is OK and not too aggressive.

Finally, we hyperlink to Bugzilla. If we're in GitHub mode, we engage a more
conservative regular expression to find bug references. This expression
does not match the "#\d+" syntax.

There are definitely some edge cases in the code. For example, if a source
repo has "#\d+" in it but it isn't GitHub, "#\d+" will appear both inside
an "href" attribute in an <a> and as part of the text value of that <a>
*then* it will get expanded again by the Bugzilla regular expression
substitution. This will obviously render incorrectly. I'm debating
whether I care enough to fix it.

IMO "perfect is the enemy of done" applies to this code. The important
thing is that the function is compatible with existing behavior on
hg.mozilla.org in cases where the source repository annotation is
missing. Also, it is important that the generated HTML is escaped:
the content comes from users and can't be trusted. Honestly, I'm not
sure if cgi.escape() is adequate.

MozReview-Commit-ID: FoVxQvS7U22