hgmo: rename "buglink" to "mozlink" (bug 1328717); r=wlach
The filter no longer links to just Bugzilla: it now links to many
things. So the name "buglink" is no longer appropriate. Plus,
Mercurial has a built-in "buglink" special variable in extensions,
that makes searching for occurrences of the filter harder than it
should be.

Rename the filter to "mozlink" for clarity.

MozReview-Commit-ID: 8PGxMidlkeb