hghooks: allow hgrc to whitelist new roots
We plan to add Servo to mozilla-central in a new root. The existing
single root hook would prevent this.

In this commit, we add functionality to the single root hook
to allow new roots if they are whitelisted in the config file.

See the inline comment for rationale of the design.