vcssync: support for converting a linearized Git repo to Mercurial (bug 1322769); r=glob
At Mozilla, we like monolithic repositories ("monorepos") where all the
code lives in one repository. This has a number of advantages,
especially around workflows and productivity. (We share this opinion with
Google and Facebook among others.)

At Mozilla, we also like open source. And a lot of open source happens
in Git and on GitHub.

Over the years, a number of groups at Mozilla have either started doing
development work on Git[Hub] or have wanted to. This creates a problem
for managing the Firefox source repository because a) it is canonically
hosted in Mercurial b) it is a monorepo. The existence of Git-based
development outside of mozilla-central fragments our code and workflows
and creates an impedance mismatch between version control tools. While
there may be local wins, overall it creates chaos.

Our solution to this problem is to provide the best of both worlds:
allow people to use Git[Hub] if they want while still providing all the
benefits of the Mercurial-backed monorepo.

This commit introduces code for converting a "linearized" Git repo to
a Mercurial one. It builds on top of the previous commit implementing
in-place Git history rewriting and adds `hg convert` to convert the
rewritten Git history to Mercurial. The code is really pretty simple.
Most of the lines deal with argument/parameter passing. The hard work
is done elsewhere.

The functionality as implemented is sufficient for a one-time or ongoing
conversion of a Git repo to Mercurial. However, what it doesn't yet do
(which will be required for ongoing vendoring of Servo commits), is
"overlaying" the Git-based files onto the Mercurial repository. This
isn't important for the "stage 1" import of Servo into mozilla-central
so it will be implemented in another commit.

MozReview-Commit-ID: HcF3LIrQYZV