Bug 1331635 - error building docker bmoweb image. r=mars,mcote
MySQL in the bmoweb docker container refuses to use
the my.cnf because it has too free permissions, causing the
container to crash since it depends on the db. This commit
prevents the world from writing to my.cnf, allowing MySQL to
come up correctly and the container code to execute properly.

Also edits the bmoweb builder entrypoint to be more descriptive
in the error message it outputs.

MozReview-Commit-ID: 8mdn2NmJIbz