mozreview: pass full commit context when rendering commit summary table fragment (bug 1330205). r=davidwalsh,smacleod
Originally, the way the commit table context was being created on
page load differed from the way it was created on a partial page
refresh. The partial page code was incomplete. Instead of duplicating
the full code, I moved out the code for getting the context into
a helper function placed in a new file called `template_helpers.py`.
Now in both cases the commit table should be rendered consistently.

MozReview-Commit-ID: 9AETsufylCh