vcssync: support for linearizing a Git repo (bug 1322769); r=glob
At Mozilla, we like linear history. It is simple and easy to reason
about. Contrast with history containing lots of merges, which can
sometimes look like a spiderweb in graphical form.

Unfortunately, the world (and arguably many tools) do not share
our rather strong opinion that linear history is preferred.

Git and GitHub are rife with non-linear history. Standard Git
workflows rely heavily on merge commits and shy away from history
rewriting. GitHub reinforces this with its default "Merge Pull Request"
workflow, which takes commits on a branch and merges them into the
destination repository. It wasn't until a few months ago that GitHub
added a mechanism to rebase (not merge) commits to facilitate a linear
history. By then, the damage was done: nearly every repository active
on GitHub contains an ugly cobweb of history with merge commits galore.

It is desired to import the Servo project into mozilla-central with
history. The Servo Git repository relies on merge commits. Although
in fairness, its history is *very* clean compared to most Git[Hub]
projects. Pretty much every commit in the first-parent ancestry of
the repository is a merge commit for a pull request and the
non-first-parent ancestry has very few merge commits. But as
relatively clean as the history is, it still isn't linear.

Furthermore, importing the non-first-parent ancestry of Servo into
mozilla-central would violate mozilla-central's goal of being
commit-level bisectable. This means that every commit in the repo can
build and can be tested. Unfortunately, the prevalent Git[Hub] commit
authoring model doesn't necessarily rely on this. There are often tons
of e.g. "fixup" commits. I like to ca;; these "how the sausage is
made" commits. These break commit-level bisection and provide little
value in the context of mozilla-central.

With that bit of history lesson out of the way, this commit introduces
functionality for "linearizing" a Git repository. You give it a Git
repo URL and ref and it iterates through the first parent ancestry of
that ref and rewrites commits so they are not merges. It stores the
rewritten commits alongside the original, tracked by a different ref.

As part of supporting the conversion of the Servo repository, this
rewriting also has features that will facilitate import of the
commits into mozilla-central will minimal additional rewriting. These
features include:

* Ability to exclude directories from history (Servo has 100,000+ WPT
  files which we don't wish to import... yet).
* Ability to prefix the commit message with a string.
* Ability to annotate the commit message with the original source
  repo location and commit.
* Ability to rewrite Reviewable.io Markdown boilerplate into something
  more concise.

The code was written in support of vendoring Servo. But it can be used
against *any* Git repository and mostly "just works." There are parts of
the code that could likely be abstracted or extended a bit better, such
as commit message rewriting (that Reviewable.io rewriting has no place
in a generic function IMO).

A lot of work went into making this code fast because tweaking things and
waiting several minutes to see results was extremely frustrating. Initial
attempts at using `git filter-branch` yielded 4+ hour execution times. I
got this down to ~2 hours by optimizing things as much as I could figure
out how. After profiling and realizing most of the time was spent in Git
index operations, I bailed on the approach and rewrote the rewriting in
Python using Dulwich. The epic is described in /docs/vcssync.rst. The
tl;dr is the Python+Dulwich approach takes ~10s instead of ~2 hours.
(Yes, you read that correctly.)


MozReview-Commit-ID: z7lwEGXYfv