ansible/hg-web: systemd units to manage repos bind mount (bug 1329738); r=fubar
Currently, we had to run an Ansible playbook after rebooting an hgweb
machine in order to mount the hg repos bind mount, which is used by
moz.build sandbox evaluation.

This is because we never took the time to figure out how to get a
read-only bind mount to work.

Getting a read-only bind mount to work is surprisingly difficult. Until
very recently, `mount` didn't support -o bind,ro: you had to do a bind
mount then `mount -o remount,bind,ro` to get a read-only bind mount.
Unfortunately, the version of mount on CentOS 7 doesn't yet support this
syntax.

A workaround is to have 2 lines in fstab, one with -o bind and a
subsequent with -o remount,bind,ro. Unfortunately, this doesn't
work with systemd! History of this is documented at
https://blog.iwakd.de/systemd-fstab-and-bind-mounts-with-options.

So, we have to resort to hacks.

We introduce a systemd mount unit that establishes a bind mount. That
is simple enough. But, the mount is read-write.

We also introduce a oneshot service unit that depends on the mount
unit. This service invokes `mount -o remount,bind,ro` to remount the
bind mount.

While we're here, we also configure the new service unit to create the
tmpfs mount for /dev inside the mount and to create a urandom character
file. Previously, this was done in Ansible and required manual action
after reboots.

With this change, the mount should come fully online after a reboot
without any Ansible invocation needed.

MozReview-Commit-ID: EpppLs9Z0YR