ansible: globally upgrade mod_wsgi from 4.4.11 to 4.5.14 (bug 1333978); r=fubar
The new version of mod_wsgi appears to just work as a drop-in
replacement. I diffed the C code and not much has changed in the
past 1+ year. There are some new features we could take advantage
of (see bug). But I'd like to get this deployed and running before
we tweak settings.

MozReview-Commit-ID: 9FL5NMqoNAW