ansible/hg-ssh: monitor master.%H file and reconcile service state (bug 1278643); r=fubar
There exists a /repo/hg/master.%H file on the NFS volume indicating
which hgssh server is the master. If this file isn't present,
hg-master.target and associated units refuse to start.

Before this commit, this file only influenced initial service
starting. If the file was deleted, services on the (old) master would
continue running. This could lead to services running on multiple
servers, which would be bad.

In this commit, we add a systemd timer that fires every few seconds.
It runs a simple script that looks at the status of hg-master.target
and the master.%H file and reconciles the active state of
hg-master.target as needed.

The solution isn't perfect. For example, there are race conditions
between deleting a file and creating a new file for a separate
master. And, multiple master.%H files may exist. But, this is a step
in the right direction.

MozReview-Commit-ID: HtY0xZ0Bj6n