vcssync: drop commits without tree changes when linearizing (bug 1322769); r=glob
Currently, the linearizing of a Git repo could result in "empty"
commits. These are commits where there are no tree changes from the
previous commit. This can happen when a rewritten commit only changes
files in directories that are filtered out as part of rewriting.

When these empty commits are fed into `hg convert`, they are silently
ignored. This makes the log output from conversion a bit confusing,
as you'll e.g. see 8 Git commits being rewritten but only 5 new
Mercurial changesets.

In this commit, we drop empty commits when linearizing Git repos.
Dropped commits are logged as such to make it easier to understand
what's going on.

MozReview-Commit-ID: auwRtt8KON