hgserver: support generating zstd bundles (bug 1331084); r=glob
This commit introduces *optional* support for generating Mercurial
bundles compressed with the zstd compression engine.

Only "version 2" bundles support zstd compression. So, there is only
a "zstd-v2" bundle type defined. This required the introduction of a
new "all" option to allow generation of these bundles on both
generaldelta and non-generaldelta repos, since the only behavior was
either or.

There are 2 zstd bundle flavors defined: a default one and a "max"
setting that cranks up the compression as far as it can go. There are
separate options for both and they are both disabled by default. This
is because I wish to be conservative with the initial zstd rollout
and only enable it on specific repositories. Eventually, we'll enable
zstd bundle generation by default and only enable zstd-max for high
volume repos. This is because zstd level 22 takes ~45m to run against
mozilla-central compared to ~7m for the default level 3. We do want
to support both because the size difference between level 3 and level
22 for mozilla-central is ~494MB (1,217MB versus ~723MB)!

As the lack of changes to existing tests demonstrate, each repo needs
to opt in to producing zstd bundles. And, the added tests show that
zstd bundles are produced as expected.

MozReview-Commit-ID: 7c2WZW1aotu