mozreview: send a series of commands to SQLite running in docker container (bug 1248008) r=smacleod
Sending a series of commands to SQLite is a complicated process, it involves creating a file, copying it over to docker container and finally ordering SQLite to read from it. This patch is creating a `mozreview sql` mach command to make the process simple to use within integration tests.

Command reads from STDIN and creates a temporary file which is copied to docker container. After SQLite finishes, all created files are removed.

MozReview-Commit-ID: Gaj9SnHd9Jf