ansible/hg-ssh: create a virtualenv with Mercurial 4.1 (bug 1331084); r=glob
Mercurial 4.1 supports zstd compression. We want to roll out zstd
bundles for clone bundles ASAP because it is a compelling performance
win. However, our policy is we don't roll out new releases of Mercurial
until a point release is available (usually after a month). This policy
stems from extra caution so we don't risk the stability of
hg.mozilla.org (we're mostly concerned about memory leaks and
performance regressions which may slip into the first major release).

These concerns don't apply as much to bundle generation. And IMO
the benefits of zstd bundles (drastically smaller sizes) outweigh
the marginal risk associated with using a .0 release before the point
release.

So, this commit introduces a virtualenv on the hgssh machines with
Mercurial 4.1.

MozReview-Commit-ID: ApC6ROR0Y3I