vcssync: daemon to trigger Servo events in response to Pulse (bug 1337572); r=glob
This commit introduces a daemon that listens for GitHub and hg.mo
Pulse events and triggers Servo-related VCS syncing actions. Well,
that's the theory. Currently, it doesn't actually trigger anything.
For that, we need Ansible infrastructure in place to define what
is triggered. That will be implemented in a later commit.

MozReview-Commit-ID: 4RQGEJiRjyK