vcssync: command for performing hg repo overlaying (bug 1331697); r=glob
Now that we have a low-level `hg overlay` command, we need to put it
to use.

This commit introduces a Python API and CLI program for invoking
`hg overlay`. It lives at a layer above `hg overlay` and provides
functionality useful for the VCS syncing service. Although, the added
value seems somewhat minimal: I'm halfway tempted to roll the
functionality into `hg overlay`. The reason I didn't is because I
originally envisioned `hg overlay` being a low-level "plumbing"
command and not a core part of a version control "syncing" tool
per se.

Anyway, the added routine is basically a specialized version of
Autoland. You give it URLs to repos, the revision and sub-directory
in the destination for the overlay, and it does the rest. It can
even push the result when it is done.

`hg overlay` has no persistent state in the destination repository.
Instead, it uses changeset metadata to track what was processed.
This means that the overlay is mostly deterministic (only thing
that should change is parents the overlay is operated on). The
functionality in this commit relies on that mostly-deterministic
behavior. One of the first things done after invocation is to
`hg strip` mutable changesets. So if there were changesets from
a previous `hg overlay` sitting around, they are wiped away
and the operation is performed again. This conveniently means
that intermittent failures due to tree closures, push races, etc
just need to be retried and things should recover. Some of this
retry logic is built into the CLI interface. It isn't as robust
as Autoland. But I think that's OK for a first iteration.

MozReview-Commit-ID: KJi6b0AXVSp