ansible: role to manage version control syncing services (bug 1337626); r=glob, fubar
Introduced is an Ansible role for a machine performing version
control "syncing" functionality. The idea is for all our services
doing version control sync to be defined in this role. Currently
implemented is configuration required for Servo VCS syncing.

I still need to write up docs for this. But this is largely working
in the real world.

MozReview-Commit-ID: Ks0T2DvOsOJ