vcssync: support uploading SHA-1 conversion map to S3
If the .hg/shamap file is lost, the Git -> Hg conversion will need to
be performed again. If this doesn't run on the *exact* configuration
that previous results were generated from, the output may diverge.

We help prevent this by backing up the shamap file to S3 and
restoring it during system provisioning.

The implementation is somewhat hacky, as it relies on the `aws` CLI
tool being installed and credentials being found by the environment.
But I was pressed for time today...

MozReview-Commit-ID: JVap7czZ59g